fix: moved files into more appropriate locations
