+++ app/src/workflow/WorkflowEngineHeaderBackButton.tsx
@@ -1,14 +1,14 @@
-import { useWorkflowEngine } from '@/contexts/WorkflowEngineContext'
 import { HeaderBackButton, HeaderBackButtonProps } from '@react-navigation/elements'
 import React from 'react'
+import { useWorkflowEngine } from './useWorkflowEngine'
 
 /**
- * Creates a custom back button for onboarding headers that navigates to the previous step in the workflow.
+ * Creates a custom back button for workflow engine headers that navigates to the previous step in the workflow.
  *
  * Note: This is a curried function to avoid re-rendering in navigation stacks.
  *
  */
-export const createOnboardingHeaderBackButton = () => {
+export const createWorkflowEngineBackHeaderButton = () => {
   // Declared so that it has a display name for debugging purposes
   const HeaderLeft = (props: HeaderBackButtonProps) => {
     const workflowEngine = useWorkflowEngine()
@@ -17,7 +17,11 @@ export const createOnboardingHeaderBackButton = () => {
       <HeaderBackButton
         {...props}
         onPress={() => {
-          workflowEngine.previousStep()
+          if (workflowEngine.currentStep.previousStep == null) {
+            throw new Error('No previous step defined in the workflow.')
+          }
+
+          workflowEngine.goToStep(workflowEngine.currentStep.previousStep)
         }}
       />
     )
